requirements.txt
opendartreader
pykrx
pandas
reportlab
python-dotenv
yfinance
beautifulsoup4
requests
streamlit
